Added some new features to classSpecies.f90. It now gives particle properties for different species sorted to the same order as they are configured in salsa namelist (and as they are in the model mass arrays). Also, finished implementing classSpecies.f90 throughout the code and made some associated simplifications to the code (a lot more can be done).
Other major additions: In SALSA all the aerosol information for all
particle categories are now stored in allSALSA-array, which is declared
in mo_salsa_driver. Operations, which are identical for all particles,
can now be handled in single loops. Category-specific access to
concentrations is handled by pointer arrays "aero", "cloud" etc., which
work exactly as the separate mass arrays before. These are declared in
mo_submctl, from where they should be imported to other modules. This
reduces the number of procedure arguments. Some simplifications making
use of this have been implemented, but a lot more can (and should) be
done in this case as well.

Further, SALSA NO LONGER DEFINES SPECIES THAT ARE NOT CONFIGURED ACTIVE
IN THE NAMELIST. Therefore, HARDCODED SPECIES INDICES CANNOT BE USED
ANYMORE. The size of the volume concentration array is still
specified as 8, since for some technical reasons this cannot be
allocatable. Will look into this later, for now it doesn't really
matter.

Fixed a bug in mo_salsa_properties, where in the calculation of dwet
the power 1/3 for HN/NO volume was in the wrong place. Luckily, this
hasn't caused problems for any "production" runs, since the compounds
haven't been used.

In this commit, the code should compile, but the major changes haven't
been tested well because atm. Voima is out of use.