Updates to ice formation and statistics.
Updated ice and snow mass concentration limits for statistics and cleaning ghost particles.
Nudging could be based on the initial soundings instead of the possibly perturbed profiles,
but this is currently not the default. Added a new versatile function for generating ice
with fixed ice number concentration. Also, prognostic ice nucleation is modified so that
ice can use snow bins. Currently these are based on hard-coded options.

 On branch IceDevel
 Changes to be committed:
	modified:   ../src/src_LES/stat.f90
	modified:   ../src/src_LES/step.f90
	modified:   ../src/src_salsa/mo_salsa.f90
	modified:   ../src/src_salsa/mo_salsa_cloud.f90