Updates related to coagulation and other minor changes.
Coagulation is modified so that this it can be updated later. For this,
cloud activation is modified so that all newly formed cloud droplets
can be described with rain bins. Also, ice and snow water in SALSA are
now described with liquid water volume. Ice and snow densities are not
used anymore to calculate particle sizes in SALSA or LES. Added an
option for LES for excluding prognostic clouds.

Replaced the original SALSA distr_update with one where update
conditions can be adjusted. Also changed the default values so that
updates are less frequent. Condensation time step is now constant
0.02 s.Added level 3 cloud and rain droplet size and concentration
outputs, and reduced the rain water threshold.

 On branch IceOrg_dev
 Your branch is up to date with 'origin/IceOrg_dev'.

 Changes to be committed:
	modified:   ../src/src_LES/LES.f90
	modified:   ../src/src_LES/grid.f90
	modified:   ../src/src_LES/ncio.f90
	modified:   ../src/src_LES/stat.f90
	modified:   ../src/src_LES/step.f90
	modified:   ../src/src_VBS/mo_vbs_partition.f90
	modified:   ../src/src_salsa/mo_salsa.f90
	modified:   ../src/src_salsa/mo_salsa_cloud.f90
	modified:   ../src/src_salsa/mo_salsa_driver.f90
	modified:   ../src/src_salsa/mo_salsa_dynamics.f90
	modified:   ../src/src_salsa/mo_salsa_init.f90
	modified:   ../src/src_salsa/mo_salsa_update.f90
	modified:   ../src/src_salsa/mo_submctl.f90